RepoRover/requirements.txt
requests==2.31.0
PyGithub==2.1.1
customtkinter==5.2.0
pillow==10.0.0 